+++ test/files/file-list.js
@@ -71,6 +71,13 @@ module.exports = [
         ]
     },
     {
+        category: "Column",
+        files: [
+            {file: "column/column-fill_balance.html", title: "column-fill: balance"},
+            {file: "column/column-fill_balance-all.html", title: "column-fill: balance-all"}
+        ]
+    },
+    {
         category: "Table",
         files: [
             {file: "table/table_at_flow_root.html", title: "Table at flow root"},